old-tests/submission/W3C/domhtml/HTMLDocument14.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/dom.html#dom-document-getelementsbyname" rel="help">
<link rel="author" title="NIST">
</head>
<body id="TEST-BODY"> 
<h1>This is a Document</h1>
<p>This is a paragraph.</p>
<img alt="Yellow image" height="50" src="/support/images/yellow-rectangle.png" width="100">
<p class="copyright"><a href="http://www.csail.mit.edu/"><acronym title="Massachusetts Institute of Technology">MIT</acronym></a>, <a href="http://www.ercim.eu/"><acronym title="European Research Consortium for Informatics and Mathematics">ERCIM</acronym></a>,
<a href="http://www.keio.ac.jp/">Keio</a></p>


<script>
var domtest = async_test("The getElementsByName method returns the (possibly empty) collection     of elements whose name value is given by the elementName.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByName("noid");
      assert_equals(nodeList.length,0);

    domtest.done();
}));

</script>
<p>The getElementsByName method returns the (possibly empty) collection
    of elements whose name value is given by the elementName.  

    Retrieve all the elements whose name attribute is "noid".
    Check the length of the nodelist.  It should be 0 since
    the id "noid" does not exist.</p>
<div id="log"></div>
</body></html>
